introduced ActiveAndroid sqlite wrapper. user add, remove works

+++ app/src/main/java/com/mk/familyweighttracker/Models/UserReading.java
@@ -6,13 +6,13 @@ import java.util.Date;
  * Created by mvalhekar on 26-03-2016.
  */
 public class UserReading {
-    public final int UserId;
+    public final long UserId;
     public final int Sequence;
     public final double Weight;
     public final double Height;
     public final Date TakenOn;
 
-    public UserReading(int userId, int sequence, double weight, double height, Date takenOn)
+    public UserReading(long userId, int sequence, double weight, double height, Date takenOn)
     {
         UserId = userId;
         Sequence = sequence;
